refac: move some files to `zk/plonk/base`
- `ColumnKey`
- `ColumnType`
- `TableXXX`
- `Phase`
