+++ tachyon/zk/expressions/evaluator/BUILD.bazel
@@ -30,7 +30,7 @@ tachyon_cc_library(
         "//tachyon/zk/expressions:scaled_expression",
         "//tachyon/zk/expressions:selector_expression",
         "//tachyon/zk/expressions:sum_expression",
-        "//tachyon/zk/plonk/circuit:ref_table",
+        "//tachyon/zk/plonk/base:ref_table",
     ],
 )
