+++ tachyon/zk/plonk/base/column_type_stringifier.h
@@ -4,13 +4,13 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_COLUMN_TYPE_STRINGIFIER_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_COLUMN_TYPE_STRINGIFIER_H_
+#ifndef TACHYON_ZK_PLONK_BASE_COLUMN_TYPE_STRINGIFIER_H_
+#define TACHYON_ZK_PLONK_BASE_COLUMN_TYPE_STRINGIFIER_H_
 
 #include <ostream>
 
 #include "tachyon/base/strings/rust_stringifier.h"
-#include "tachyon/zk/plonk/circuit/column_type.h"
+#include "tachyon/zk/plonk/base/column_type.h"
 
 namespace tachyon::base::internal {
 
@@ -25,4 +25,4 @@ class RustDebugStringifier<zk::ColumnType> {
 
 }  // namespace tachyon::base::internal
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_COLUMN_TYPE_STRINGIFIER_H_
+#endif  // TACHYON_ZK_PLONK_BASE_COLUMN_TYPE_STRINGIFIER_H_
