+++ tachyon/zk/plonk/base/ref_table.h
@@ -4,12 +4,12 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_REF_TABLE_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_REF_TABLE_H_
+#ifndef TACHYON_ZK_PLONK_BASE_REF_TABLE_H_
+#define TACHYON_ZK_PLONK_BASE_REF_TABLE_H_
 
 #include <vector>
 
-#include "tachyon/zk/plonk/circuit/table_base.h"
+#include "tachyon/zk/plonk/base/table_base.h"
 
 namespace tachyon::zk {
 
@@ -43,4 +43,4 @@ class RefTable : public TableBase<PolyOrEvals> {
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_REF_TABLE_H_
+#endif  // TACHYON_ZK_PLONK_BASE_REF_TABLE_H_
