+++ tachyon/zk/plonk/base/ref_table_unittest.cc
@@ -1,4 +1,4 @@
-#include "tachyon/zk/plonk/circuit/ref_table.h"
+#include "tachyon/zk/plonk/base/ref_table.h"
 
 #include <vector>
