+++ tachyon/zk/plonk/circuit/assignment.h
@@ -14,7 +14,7 @@
 #include "tachyon/math/base/rational_field.h"
 #include "tachyon/zk/base/row_index.h"
 #include "tachyon/zk/base/value.h"
-#include "tachyon/zk/plonk/circuit/column_key.h"
+#include "tachyon/zk/plonk/base/column_key.h"
 #include "tachyon/zk/plonk/constraint_system/challenge.h"
 #include "tachyon/zk/plonk/constraint_system/selector.h"
