+++ tachyon/zk/plonk/circuit/cell.h
@@ -15,7 +15,7 @@
 
 #include "tachyon/export.h"
 #include "tachyon/zk/base/row_index.h"
-#include "tachyon/zk/plonk/circuit/column_key.h"
+#include "tachyon/zk/plonk/base/column_key.h"
 
 namespace tachyon::zk {
