+++ tachyon/zk/plonk/circuit/floor_planner/v1/BUILD.bazel
@@ -7,8 +7,8 @@ tachyon_cc_library(
     hdrs = ["assignment_pass.h"],
     deps = [
         ":v1_plan",
+        "//tachyon/zk/plonk/base:column_key",
         "//tachyon/zk/plonk/circuit:cell",
-        "//tachyon/zk/plonk/circuit:column_key",
         "//tachyon/zk/plonk/circuit:layouter",
         "//tachyon/zk/plonk/circuit/floor_planner:lookup_table_assigner",
         "//tachyon/zk/plonk/circuit/floor_planner:plan_region",
@@ -69,8 +69,8 @@ tachyon_cc_library(
         "//tachyon:export",
         "//tachyon/base:logging",
         "//tachyon/base/containers:container_util",
-        "//tachyon/zk/plonk/circuit:column_key",
-        "//tachyon/zk/plonk/circuit:column_type",
+        "//tachyon/zk/plonk/base:column_key",
+        "//tachyon/zk/plonk/base:column_type",
         "//tachyon/zk/plonk/circuit:region_shape",
         "//tachyon/zk/plonk/circuit/floor_planner:allocations",
         "@com_google_absl//absl/container:flat_hash_map",
