+++ tachyon/zk/plonk/circuit/floor_planner/v1/assignment_pass.h
@@ -12,8 +12,8 @@
 #include <string_view>
 #include <utility>
 
+#include "tachyon/zk/plonk/base/column_key.h"
 #include "tachyon/zk/plonk/circuit/cell.h"
-#include "tachyon/zk/plonk/circuit/column_key.h"
 #include "tachyon/zk/plonk/circuit/floor_planner/lookup_table_assigner.h"
 #include "tachyon/zk/plonk/circuit/floor_planner/plan_region.h"
 #include "tachyon/zk/plonk/circuit/floor_planner/scoped_region.h"
