+++ tachyon/zk/plonk/circuit/lookup_table_column.h
@@ -12,7 +12,7 @@
 #include "absl/hash/hash.h"
 
 #include "tachyon/export.h"
-#include "tachyon/zk/plonk/circuit/column_key.h"
+#include "tachyon/zk/plonk/base/column_key.h"
 
 namespace tachyon::zk {
