+++ tachyon/zk/plonk/circuit/region_layouter.h
@@ -11,8 +11,8 @@
 #include "tachyon/math/base/rational_field.h"
 #include "tachyon/zk/base/row_index.h"
 #include "tachyon/zk/base/value.h"
+#include "tachyon/zk/plonk/base/column_key.h"
 #include "tachyon/zk/plonk/circuit/assigned_cell.h"
-#include "tachyon/zk/plonk/circuit/column_key.h"
 #include "tachyon/zk/plonk/constraint_system/selector.h"
 
 namespace tachyon::zk {
