+++ tachyon/zk/plonk/constraint_system/query.h
@@ -11,7 +11,7 @@
 
 #include "absl/strings/substitute.h"
 
-#include "tachyon/zk/plonk/circuit/column_key.h"
+#include "tachyon/zk/plonk/base/column_key.h"
 #include "tachyon/zk/plonk/constraint_system/rotation.h"
 
 namespace tachyon::zk {
