+++ tachyon/zk/plonk/halo2/pinned_constraint_system.h
@@ -12,7 +12,7 @@
 #include <vector>
 
 #include "tachyon/zk/lookup/lookup_argument_stringifier.h"
-#include "tachyon/zk/plonk/circuit/phase_stringifier.h"
+#include "tachyon/zk/plonk/base/phase_stringifier.h"
 #include "tachyon/zk/plonk/constraint_system/constraint_system.h"
 #include "tachyon/zk/plonk/constraint_system/query_stringifier.h"
 #include "tachyon/zk/plonk/halo2/pinned_gates.h"
