+++ tachyon/zk/plonk/permutation/permutation_argument.h
@@ -12,7 +12,7 @@
 
 #include "tachyon/base/containers/contains.h"
 #include "tachyon/export.h"
-#include "tachyon/zk/plonk/circuit/column_key.h"
+#include "tachyon/zk/plonk/base/column_key.h"
 
 namespace tachyon::zk {
