+++ tachyon/zk/plonk/permutation/permutation_argument_stringifier.h
@@ -10,7 +10,7 @@
 #include <ostream>
 
 #include "tachyon/base/strings/rust_stringifier.h"
-#include "tachyon/zk/plonk/circuit/column_key_stringifier.h"
+#include "tachyon/zk/plonk/base/column_key_stringifier.h"
 #include "tachyon/zk/plonk/permutation/permutation_argument.h"
 
 namespace tachyon::base::internal {
