+++ vendors/halo2/BUILD.bazel
@@ -226,8 +226,8 @@ tachyon_cc_library(
         "//tachyon/math/polynomials/univariate:univariate_evaluations",
         "//tachyon/math/polynomials/univariate:univariate_polynomial",
         "//tachyon/zk/base/commitments:shplonk_extension",
-        "//tachyon/zk/plonk/circuit:column_key",
-        "//tachyon/zk/plonk/circuit:phase",
+        "//tachyon/zk/plonk/base:column_key",
+        "//tachyon/zk/plonk/base:phase",
         "//tachyon/zk/plonk/constraint_system:gate",
         "//tachyon/zk/plonk/halo2:pinned_verifying_key",
         "//tachyon/zk/plonk/keys:proving_key",
